Moved kernel to force full 64-bit support for things like integer wme values, timetags, stats and counters, rete data structures.
General notes:
- Motivation for this is support for long-running agents. We're already rolling over stats numbers and timetags.
- Arguably this should be on a branch, but it is easier to check in and force testing by just putting it on the trunk. I'll roll it back and move it to a branch if there are problems or sufficient resistance.
- This check-in only compiles on and is tested on Windows. Fixes for linux and mac will be in the next couple check-ins.
- The "long long" workaround for SWIG (see below) may not work right on linux/mac and, if it doesn't, I'll roll back this change off the trunk and put it on a branch until we figure out a better workaround for SWIG and int64_t.

Details:
- Removal of hackish usage of intptr_t and uintptr_t as a flexible-size integer field. Removed associated data structures in soar_db.
- "long" and "unsigned long" are 64-bits on GCC on linux, mac; only 32-bits on MSVC. Changed these fields to the C99 stdint.h standard int64_t and uint64_t throughout the kernel and SML so that all platforms now see them as 64-bits. This could potentially seriously impact performance on 32-bit platforms.
- SWIG (even the newest version) currently interprets int64_t as some kind of pointer type on the hardware I've tested it on so far. In order to fix this, I changed instances of int64_t to "long long" which it correctly converts. This might not work on GCC.
- There are still plenty of 32-bit integers in the kernel and SML, the main changes were to stats, int wme values, timetags, rete data structures, tc numbers.
- The ClientSML interface was added to so that the change is backwards-compatible when compiled with reasonable warning levels. New functions have been added for those that need to return 64-bit values and are incapable of overloading (such as IntElement::GetValue() which takes no arguments). These functions are named the same with a 64 appended to the name (IntElement::GetValue64()). 
- Rete net loading and saving updated so that the new 64-bit rete data structures can be saved with full precision. This is implemented as a new rete net file version (4 instead of 3). The old 32-bit rete-nets (version 3) can still be loaded with no changes. The rete-net command should probably take a new argument to enable 32-bit saves (loads are auto-detected). I've tested this locally but have not made a cppunit test for it yet.
- The random number generator is 32-bit, but was changed to use stdint.h instead of it's own 32-bit type.
- There is no 64-bit version of the SML tag for the integer type. These are used when passing results and arguments and such back and forth through SML. These integers can be referenced as 32-bits or 64-bits as necessary, they are stored in signed 64-bit fields on the structures or as strings of arbitrary precision when rendered to XML.

+++ Core/CLI/src/cli_CommandLineInterface.h
@@ -29,8 +29,8 @@
 #include "cli_Aliases.h"
 #include "kernel.h"
 
-typedef uintptr_t epmem_time_id;
-typedef uintptr_t smem_lti_id;
+typedef uint64_t epmem_time_id;
+typedef uint64_t smem_lti_id;
 
 // For test
 //#define WIN_STATIC_LINK
@@ -303,7 +303,7 @@ public:
 	* @param pPrefix Pointer to the new prefix, must not contain '*' character, 
 	*        null for query
 	*************************************************************/
-	bool DoChunkNameFormat(const bool* pLongFormat = 0, const int* pCount = 0, const std::string* pPrefix = 0);
+	bool DoChunkNameFormat(const bool* pLongFormat = 0, const int64_t* pCount = 0, const std::string* pPrefix = 0);
 
 	/*************************************************************
 	* @brief clog command
@@ -396,7 +396,7 @@ public:
 	* @brief gp-max command
 	* @param maximum The maximum number of productions to allow generation of
 	*************************************************************/
-	bool DoGPMax(const long& maximum);
+	bool DoGPMax(const int& maximum);
 
 	/*************************************************************
 	* @brief help command
@@ -582,7 +582,7 @@ public:
 	* @brief remove-wme command
 	* @param timetag The timetag of the wme to remove
 	*************************************************************/
-	bool DoRemoveWME(unsigned long timetag);
+	bool DoRemoveWME(uint64_t timetag);
 
 	/*************************************************************
 	* @brief replay-input command
@@ -646,7 +646,7 @@ public:
 	* @param pAttr the attribute to get/set/stats, pass 0 (null) only if no pOp (all config) or stats (full stats)
 	* @param pVal the value to set, pass 0 (null) only if no pOp (all config), get, or stats
 	*************************************************************/
-	bool DoSMem(const char pOp = 0, const std::string *pAttr = 0, const std::string *pVal = 0, smem_lti_id lti_id = 0, unsigned long depth = 0);
+	bool DoSMem(const char pOp = 0, const std::string *pAttr = 0, const std::string *pVal = 0, smem_lti_id lti_id = 0, unsigned int depth = 0);
 
 	/*************************************************************
 	* @brief soar8 command
@@ -676,7 +676,7 @@ public:
 	* @param pSeed Number to seed the random number generator with, pass
 	*		 null to seed randomly.
 	*************************************************************/
-	bool DoSRand(unsigned long int* pSeed = 0);
+	bool DoSRand(uint32_t* pSeed = 0);
 
 	/*************************************************************
 	* @brief stats command
@@ -896,7 +896,7 @@ protected:
 	bool				m_EchoResult;			// If true, copy result of command to echo event stream
 	EchoMap				m_EchoMap;				// If command appears in this map, always echo it.
 	bool				m_VarPrint;				// Used in print command to put <>'s around identifiers.
-	long				m_GPMax;				// Max number of productions to allow gp to produce
+	size_t              m_GPMax;				// Max number of productions to allow gp to produce
 
 	soarxml::XMLTrace*	m_XMLResult;			// Used to collect up XML output from commands that directly support that.
 	ElementXMLList		m_ResponseTags;			// List of tags for the response.
